e2e/requirements-e2e.txt
testcontainers==3.2.0
deprecation==2.1.0
GitPython==3.1.12
gitdb==4.0.5
smmap==3.0.4
